% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace}
\alias{ace}
\title{Average Causal Effect}
\usage{
ace(x, y)
}
\arguments{
\item{x}{a value}

\item{y}{a value}
}
\value{
a value
}
\description{
Average Causal Effect
}
\examples{
ace(1:10, 11:20)
}
