% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_data_gen.R
\name{censor_data_gen}
\alias{censor_data_gen}
\title{数据生成函数：censor_data_gen}
\usage{
censor_data_gen(
  n = 5000,
  p = 10,
  PH = TRUE,
  censor = 30,
  setting = 1,
  overlap = "strong"
)
}
\arguments{
\item{n}{样本量，默认为 5000。}

\item{p}{协变量数量，默认为 10。}

\item{PH}{是否假设比例风险模型，默认为 TRUE。}

\item{censor}{删失比例，默认为 30。}

\item{setting}{数据生成的设置，默认为 1。}

\item{overlap}{协变量之间的重叠程度，可选值为 "strong"、"medium" 和 "weak"，默认为 "strong"。}
}
\value{
返回一个数据框，包含以下列：
\itemize{
\item \code{LP0}, \code{LP1}: 线性预测器。
\item \code{lambda0}, \code{lambda1}: 比例参数。
\item \code{eta_0}, \code{eta_1}: 风险模型参数。
\item \code{p1_category}: 倾向得分类别。
\item \code{x1} 至 \code{x10}: 协变量。
\item \code{W}: 二分类变量，表示分组。
\item \code{C}: 删失时间。
\item \code{Time}: 观察到的生存时间。
\item \code{Event}: 删失指标，1 表示未删失，0 表示删失。
\item \code{T0}, \code{T1}: 潜在生存时间。
\item \code{p}: 倾向得分。
}
}
\description{
生成模拟数据用于因果推断分析，支持不同删失比例、协变量重叠程度和模型设置。
}
\examples{
# 生成默认参数的数据
data <- censor_data_gen()

# 生成弱重叠、删失比例为 20\% 的数据
data <- censor_data_gen(n = 1000, overlap = "weak", censor = 20)

# 使用非比例风险模型
data <- censor_data_gen(PH = FALSE)
}
